+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/geminiVertex/GeminiClientConfiguration.kt
@@ -1,4 +1,4 @@
-package com.github.blarc.ai.commits.intellij.plugin.settings.clients.geminiApi
+package com.github.blarc.ai.commits.intellij.plugin.settings.clients.geminiVertex
 
 import com.github.blarc.ai.commits.intellij.plugin.Icons
 import com.github.blarc.ai.commits.intellij.plugin.settings.clients.LLMClientConfiguration
@@ -9,16 +9,20 @@ import com.intellij.util.xmlb.annotations.Attribute
 import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
 import javax.swing.Icon
 
-class GeminiApiClientConfiguration : LLMClientConfiguration(
-    "GeminiApi",
-    "gemini-1.5-pro",
+// Can not rename this class because of backwards compatibility
+// with persistent component - AppSettings2
+class GeminiClientConfiguration : LLMClientConfiguration(
+    "Gemini Vertex",
+    "gemini-pro",
     "0.7"
 ) {
     @Attribute
-    var apiKey: String = "api-key"
+    var projectId: String = "project-id"
+    @Attribute
+    var location: String = "us-central1"
 
     companion object {
-        const val CLIENT_NAME = "GeminiApi"
+        const val CLIENT_NAME = "Gemini Vertex"
     }
 
     override fun getClientName(): String {
@@ -26,30 +30,32 @@ class GeminiApiClientConfiguration : LLMClientConfiguration(
     }
 
     override fun getClientIcon(): Icon {
-        return Icons.GEMINI
+        return Icons.GEMINI_VERTEX
     }
 
     override fun getSharedState(): LLMClientSharedState {
-        return GeminiApiClientSharedState.getInstance()
+        return GeminiVertexClientSharedState.getInstance()
     }
 
     override fun generateCommitMessage(commitWorkflowHandler: AbstractCommitWorkflowHandler<*, *>, commitMessage: CommitMessage, project: Project) {
-        return GeminiApiClientService.getInstance().generateCommitMessage(this, commitWorkflowHandler, commitMessage, project)
+        return GeminiVertexClientService.getInstance().generateCommitMessage(this, commitWorkflowHandler, commitMessage, project)
     }
 
     // Model names are hard-coded and do not need to be refreshed.
     override fun getRefreshModelsFunction() = null
 
     override fun clone(): LLMClientConfiguration {
-        val copy = GeminiApiClientConfiguration()
+        val copy = GeminiClientConfiguration()
         copy.id = id
         copy.name = name
         copy.modelId = modelId
         copy.temperature = temperature
-        copy.apiKey = apiKey
+        copy.projectId = projectId
+        copy.location = location
         return copy
     }
 
-    override fun panel() = GeminiApiClientPanel(this)
+    override fun panel() = GeminiVertexClientPanel(this)
+
 
 }